+++ src/core/visualization.c
@@ -348,7 +348,7 @@ void WriteNode(FILE *fp, POSITION node, int level, EDGELIST *tree) {
 	}
 
 	if(!Visited(node)) {
-		if(level != GetLevelNumber(pdata)) {
+		if((unsigned) level != GetLevelNumber(pdata)) {
 			fprintf(fp, "\t\tsubgraph cluster%llu {\n", node);
 			fprintf(fp, "\t\t\tlabel = \" \"\n");
 			fprintf(fp, "\t\t\tcolor = \"blue\"\n\t");
@@ -358,11 +358,11 @@ void WriteNode(FILE *fp, POSITION node, int level, EDGELIST *tree) {
 		}
 
 		if(node == gInitialPosition) {
-			if(level == GetLevelNumber(pdata)) {
+			if((unsigned) level == GetLevelNumber(pdata)) {
 				fprintf(fp, "\t\tsubgraph cluster%llu {\n", node);
 				fprintf(fp, "\t\t\tcolor = \"blue\"\n\t");
 			}
-			fprintf(fp, "%slabel = \"Initial Position\"\n\t", (level == GetLevelNumber(pdata)) ? "\t\t\t" : "\t\t");
+			fprintf(fp, "%slabel = \"Initial Position\"\n\t", ((unsigned) level == GetLevelNumber(pdata)) ? "\t\t\t" : "\t\t");
 		}
 
 		if(GetLevelNumber(pdata) == 0) {
@@ -375,7 +375,7 @@ void WriteNode(FILE *fp, POSITION node, int level, EDGELIST *tree) {
 			BadElse("WriteNode");
 		}
 
-		if(level != GetLevelNumber(pdata) || node == gInitialPosition) {
+		if((unsigned) level != GetLevelNumber(pdata) || node == gInitialPosition) {
 			fprintf(fp, "\n\t\t}\n");
 		} else {
 			fprintf(fp, "\n");
@@ -392,7 +392,7 @@ void WriteNode(FILE *fp, POSITION node, int level, EDGELIST *tree) {
 		//REMOTENESS_MAX level above current level,
 		//REMOTENESS_MAX+1 level below current level
 		if(gRemotenessOrder) {
-			if(level == GetLevelNumber(pdata)) {
+			if((unsigned) level == GetLevelNumber(pdata)) {
 				if(GetFringe(pdata)) {
 					UpdateRankList(tree, node, REMOTENESS_MAX+1); // want fringes at bottom
 				} else {
